Add Token Policy handling
This patchset adds the policy creation and handling code as
defined in:

- https://desec.readthedocs.io/en/latest/auth/tokens.html#token-policy-management

This allows for users to manage the token policy files in-library in
order to add restrictions to individual tokens. An example use case
would be creating tokens in CI and building a policy around them to
restrict those tokens to only be able to change specific records.

Adds the following functions:

- `GetPolicies` - Gets all token policies.
- `CreatePolicy` - Creates a new token policy.
- `DeletePolicy` - Deletes a token policy.